final query handlers using indexing from DB and additional handle of erros to store todos by current user in session

+++ convex/functions.ts
@@ -1,10 +1,15 @@
 import { mutation, query } from "./_generated/server";
 import { v } from "convex/values";
+import { requireUser } from "./helper";
 
 //Function query: pulls data form DB
 export const listTodos = query({
   handler: async (ctx) => {
-    return await ctx.db.query("todos").collect(); // get all todos from DB
+    const user = await requireUser(ctx);
+    return await ctx.db
+      .query("todos")
+      .withIndex("by_user_id", (q) => q.eq("userId", user.tokenIdentifier))
+      .collect(); // get all todos from DB filtering by userId specifig such that we do not leak info to other users todos, indexing the queries
   },
 });
 
@@ -15,10 +20,12 @@ export const createTodo = mutation({
     description: v.string(),
   },
   handler: async (ctx, args) => {
+    const user = await requireUser(ctx);
     await ctx.db.insert("todos", {
       title: args.title,
       description: args.description,
       completed: false,
+      userId: user.tokenIdentifier,
     });
   },
 });
@@ -30,6 +37,11 @@ export const updateTodo = mutation({
     completed: v.boolean(),
   },
   handler: async (ctx, args) => {
+    const user = await requireUser(ctx);
+    const todo = await ctx.db.get(args.id);
+    if (todo?.userId !== user.tokenIdentifier) {
+      throw new Error("Unauthorized");
+    }
     await ctx.db.patch(args.id, {
       completed: args.completed,
     });
@@ -42,6 +54,11 @@ export const deleteTodo = mutation({
     id: v.id("todos"),
   },
   handler: async (ctx, args) => {
+    const user = await requireUser(ctx);
+    const todo = await ctx.db.get(args.id);
+    if (todo?.userId !== user.tokenIdentifier) {
+      throw new Error("Unauthorized");
+    }
     await ctx.db.delete(args.id);
   },
 });
